LayoutTests/http/tests/misc/createElementNamespace2.xhtml

<foo>
<script xmlns="http://www.w3.org/1999/xhtml" type="text/javascript" src="../security/resources/cross-frame-access.js"></script>
<pre xmlns="http://www.w3.org/1999/xhtml" id="console"></pre>
<script xmlns="http://www.w3.org/1999/xhtml" id="scriptTag">
    if (window.testRunner)
        testRunner.dumpAsText();

    log("This test was written to address rdar://problem/4976879.  And tests how namespaceURI's are assigned.");
    log("");

    log("Test1: document.namespaceURI")
    shouldBe("document.namespaceURI", "null");
    log("");

    log("Test2: document.createElement().namespaceURI")
    shouldBe("document.createElement(\"div\").namespaceURI", "'http://www.w3.org/1999/xhtml'");
    log("");

    var myScriptTag = document.getElementById("scriptTag");
    log("Test3: script tag.namespaceURI");
    shouldBe("myScriptTag.namespaceURI", "'http://www.w3.org/1999/xhtml'");
</script>
</foo>
