third_party/blink/web_tests/platform/generic/http/tests/misc/createElementNamespace2-expected.txt

This test was written to address rdar://problem/4976879.  And tests how namespaceURI's are assigned.

Test1: document.namespaceURI
PASS: document.namespaceURI should be 'undefined' and is.

Test2: document.createElement().namespaceURI
PASS: document.createElement("div").namespaceURI should be 'http://www.w3.org/1999/xhtml' and is.

Test3: script tag.namespaceURI
PASS: myScriptTag.namespaceURI should be 'http://www.w3.org/1999/xhtml' and is.
